Fixing other dialogs

+++ Editor/MenuActions/Objects/ObjectUtils.cs
@@ -99,8 +99,8 @@ namespace GameMeanMachine.Unity.WindRose
                         gameObject.transform.parent = selectedTransform;
                         gameObject.SetActive(false);
                         AlephVault.Unity.Layout.Utils.Behaviours.AddComponent<Authoring.Behaviours.Entities.Objects.MapObject>(gameObject, new Dictionary<string, object>() {
-                            { "width", (uint)objectSize.x },
-                            { "height", (uint)objectSize.y }
+                            { "width", (ushort)objectSize.x },
+                            { "height", (ushort)objectSize.y }
                         });
                         if (addCommandSender)
                         {
